-now post/update/delete of a message has authotization -changed some constuctors to primary

+++ CRUDTestProject/Controllers/MessagesController.cs
@@ -12,18 +12,10 @@ namespace CRUDTestProject.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    public class MessagesController : ControllerBase
+    public class MessagesController(IMessageRepository messageRepository) : ControllerBase
     {
-        private readonly IMessageRepository messageRepository;
-
-        public MessagesController(IMessageRepository messageRepository)
-        {
-            this.messageRepository = messageRepository;
-        }
-
-
         [HttpGet]
-        public IActionResult getMessagesPassingFilter(
+        public IActionResult GetMessagesPassingFilter(
             [FromQuery] MessageFilterParameters filterParameters,
             [FromQuery] string? matchUsername,
             [FromQuery] bool isOrderAscending = true)
@@ -36,7 +28,7 @@ namespace CRUDTestProject.Controllers
             {
                 result = result.Where(m => m.Username == matchUsername);
             }
-            result = filterParameters.filterMessages(result);
+            result = filterParameters.FilterMessages(result);
             
 
             return Ok(
@@ -48,7 +40,7 @@ namespace CRUDTestProject.Controllers
 
         [HttpGet]
         [Route("{id:guid}")]
-        public IActionResult getMessageById(Guid id)
+        public IActionResult GetMessageById(Guid id)
         {
             var message = messageRepository.GetById(id);
 
@@ -61,6 +53,7 @@ namespace CRUDTestProject.Controllers
         }
 
         [HttpPost]
+        [Authorize]
         public IActionResult AddMessage(AddMessageDto addMessageDto)
         {
             var messageEntity = new Message 
@@ -78,18 +71,32 @@ namespace CRUDTestProject.Controllers
         }
 
         [HttpPut]
+        [Authorize]
         [Route("{id:guid}")]
         public IActionResult UpdateMessage(Guid id, UpdateMessageDto updateMessageDto) 
         {
+            var posterUsername = messageRepository.GetPosterUsernameById(id);
+            
+            if (posterUsername != User.FindFirstValue(ClaimTypes.Name))
+            {
+                return Unauthorized("You can only update your own messages.");
+            }
+
             Message message = messageRepository.Update(id, updateMessageDto.Name, updateMessageDto.Content);
             
             return Ok(new MessageResponseModel(message));
         }
 
         [HttpDelete]
+        [Authorize]
         [Route("{id:guid}")]
         public IActionResult DeleteMessage(Guid id) 
         {
+            var posterUsername = messageRepository.GetPosterUsernameById(id);
+            if (posterUsername != User.FindFirstValue(ClaimTypes.Name))
+            {
+                return Unauthorized("You can delete only your own posted messages.");
+            }
             messageRepository.Delete(id);
 
             return Ok();
@@ -98,7 +105,7 @@ namespace CRUDTestProject.Controllers
         [HttpGet]
         [Authorize]
         [Route("[action]")]
-        public IActionResult getUsername()
+        public IActionResult GetUsername()
         {
             var name = User.FindFirstValue(ClaimTypes.Name);
